Re-organize IdentityProvider
