+++ package.json
@@ -53,8 +53,8 @@
     "go-ipfs": "^0.17.0",
     "ipfs": "^0.66.0",
     "ipfsd-ctl": "^13.0.0",
-    "localstorage-down": "^0.6.7",
-    "localstorage-level-migration": "^0.2.0",
+    "key-did-provider-ed25519": "^2.0.1",
+    "key-did-resolver": "^2.3.0",
     "markdown-toc": "^1.2.0",
     "mkdirp": "^2.1.1",
     "mocha": "^10.2.0",
