+++ src/index.js
@@ -1,2 +1,3 @@
 export { Log, DefaultAccessController, Entry } from './oplog/index.js'
 export { default as KeyStore } from './key-store.js'
+export { IdentityProvider } from './identities/index.js'
