+++ test/oplog/entry.test.js
@@ -2,7 +2,7 @@ import { strictEqual, deepStrictEqual } from 'assert'
 import rimraf from 'rimraf'
 import { copy } from 'fs-extra'
 import { Entry } from '../../src/oplog/index.js'
-import IdentityProvider from '../../src/identities/index.js'
+import { IdentityProvider } from '../../src/identities/index.js'
 import KeyStore from '../../src/key-store.js'
 import { config, testAPIs, startIpfs, stopIpfs } from 'orbit-db-test-utils'
 // import IdentityStorage from '../src/identity-storage.js'
