+++ test/oplog/load.test.js
@@ -3,7 +3,7 @@ import rimraf from 'rimraf'
 import { copy } from 'fs-extra'
 import { Log, Entry, Sorting } from '../../src/oplog/index.js'
 import bigLogString from '../fixtures/big-log.fixture.js'
-import IdentityProvider from '../../src/identities/index.js'
+import { IdentityProvider } from '../../src/identities/index.js'
 import KeyStore from '../../src/key-store.js'
 import LogCreator from './utils/log-creator.js'
 import MemoryStorage from '../../src/storage/memory.js'
