+++ test/oplog/signed-log.test.js
@@ -1,7 +1,7 @@
 import { notStrictEqual, strictEqual, deepStrictEqual } from 'assert'
 import rimraf from 'rimraf'
 import { Log } from '../../src/oplog/index.js'
-import IdentityProvider from '../../src/identities/index.js'
+import { IdentityProvider } from '../../src/identities/index.js'
 import KeyStore from '../../src/key-store.js'
 
 // Test utils
